refactor : 이미지 컨테이너 경로변경 #92
