+++ src/components/UserInfoWithFollow/index.tsx
@@ -1,6 +1,6 @@
-import ImgContainer from '@components/ImgContainer';
 import * as S from '@components/UserInfoWithFollow/UserInfoWithFollow.style';
 import Icon from '@components/common/Icon';
+import ImgContainer from '@components/common/ImgContainer';
 
 interface UserInfoWithFollowPropsType {
   writerThumbnailImageUrl: string;
