+++ src/components/common/ImgContainer/ImgContainer.style.tsx
